v1.2.0 + More YT Notifications tweaking
I updated the node dependencies to their latest versions.

I also changed the filename structure of the yt notification files to represent which channel it's responsible for maintaining.

The last remark is more for consistency on the main branch as the dev discord server only has my channel setup to enable debug testing there.