v1.5.0
I corrected the checkLiveTwitchPMG.js code to properly handle the notifications for my friend's twitch channel. Which was previously generating 400 errors in the console when checking for new videos since the twitch applications are linked to our individual accounts. Leading to permission errors.

Lastly, I updated the dependencies to their latest versions.